+++ src/main/java/com/springboot/insideClass/controllers/NotasController.java
@@ -1,7 +1,10 @@
 package com.springboot.insideClass.controllers;
 
-import com.springboot.insideClass.entity.*;
 import com.springboot.insideClass.componet.Correo;
+import com.springboot.insideClass.entity.AsignaturaDocenteEntity;
+import com.springboot.insideClass.entity.AsignaturaNotaEntity;
+import com.springboot.insideClass.entity.CursoEstablecimientoEntity;
+import com.springboot.insideClass.entity.MatriculaEntity;
 import com.springboot.insideClass.payload.request.Notas.CreateNotas;
 import com.springboot.insideClass.payload.request.Notas.EditNotas;
 import com.springboot.insideClass.payload.request.Notas.GetNotas;
@@ -15,7 +18,6 @@ import org.springframework.web.bind.annotation.*;
 import javax.validation.Valid;
 import java.time.LocalDate;
 import java.time.ZoneId;
-import java.util.Date;
 
 @CrossOrigin(origins = "*", maxAge = 3600)
 @RestController
@@ -42,19 +44,36 @@ public class NotasController {
     @PostMapping("/Get")
     public ResponseEntity<?> obtenerNota(@Valid @RequestBody GetNotas notaRequest) {
 
-        try{
+        System.out.println(notaRequest.getDocente_run());
+        System.out.println(notaRequest.getRun());
+        System.out.println(notaRequest.getId_asignatura());
+        System.out.println(notaRequest.getId_establecimiento());
+
+        if(asignaturaNotaService.IsValid(notaRequest)){
+
             return ResponseEntity.ok().header(HttpHeaders.SET_COOKIE)
-                    .body(asignaturaNotaService.obtenerInfoNotas(notaRequest.getRun(), notaRequest.getDocente_run(), notaRequest.getId_establecimiento(),notaRequest.getId_asignatura(), notaRequest.getFecha()));
+                    .body(asignaturaNotaService.obtenerInfoNotas(notaRequest.getRun(), notaRequest.getDocente_run(),
+                            notaRequest.getId_establecimiento(),notaRequest.getId_asignatura(), notaRequest.getFecha(), notaRequest.getCurso()));
+        }
+        return  ResponseEntity.badRequest().body(new MessageResponse("Error: No se ha logrado registrar nota!"));
+
+      /*  try{
+
         }catch (Exception e){
 
             return  ResponseEntity.badRequest().body(new MessageResponse("Error: No se ha logrado registrar nota!"));
 
-        }
+        }*/
     }
 
     @PostMapping("/Create")
     public ResponseEntity<?> ingresarNota(@Valid @RequestBody CreateNotas notaRequest) {
-      /*  try {
+       /*try {
+
+        } catch (Exception e) {
+            return ResponseEntity.badRequest().body(new MessageResponse("Error: No se ha logrado registrar la nota!" + e.getMessage()));
+        }*/
+        if(notaRequest.IsValid()){
             LocalDate localDate = notaRequest.getFecha().toInstant().atZone(ZoneId.systemDefault()).toLocalDate();
             int year = localDate.getYear();
 
@@ -63,79 +82,56 @@ public class NotasController {
                 return ResponseEntity.badRequest().body(new MessageResponse("Error: No se ha logrado encontrar matricula!"));
             }
 
-            CursoEstablecimientoEntity cursoEstablecimiento = cursoEstablecimientoService.findById(matricula.getCursoEstablecimientoEntity().getCurso_establ_id());
+
+            CursoEstablecimientoEntity cursoEstablecimiento = cursoEstablecimientoService.findCursoEstablecimientoByCursoAndEstablecimiento(-1, notaRequest.getEstablecimiento(),
+                    matricula.getMatricula_id().intValue(),
+                    year, year);
             if (cursoEstablecimiento == null) {
                 return ResponseEntity.badRequest().body(new MessageResponse("Error: No se ha logrado encontrar matricula !"));
             }
 
-            AsignaturaDocenteEntity asignaturaDocente = asignaturaDocenteService.findDocenteCursoByRunAndAsignaturaAndEstablecimiento(notaRequest.getDocente_run(), notaRequest.getAsignatura(), cursoEstablecimiento.getCursoEntity().getCurso_id(), year);
+          AsignaturaDocenteEntity asignaturaDocente = asignaturaDocenteService.findDocenteCursoByRunAndAsignaturaAndEstablecimiento(
+                  year,
+                  year,
+                  notaRequest.getDocente_run(),
+                  notaRequest.getEstablecimiento(),
+                  notaRequest.getAsignatura());
             if (asignaturaDocente == null) {
                 return ResponseEntity.badRequest().body(new MessageResponse("Error: No se ha logrado registrar nota!"));
             }
 
             // Crear nueva entidad AsignaturaNotaEntity
             AsignaturaNotaEntity nota = new AsignaturaNotaEntity();
-            nota.setMatriculaEntity(matricula);
-            nota.setAsignaturaDocenteEntity(asignaturaDocente);
             nota.setAsignatura_nota_nota(notaRequest.getNota());
-
+            nota.setPosicion_nota_nota(notaRequest.getPosicion_nota());
             // Guardar la nueva entidad en la base de datos
             asignaturaNotaService.save(nota);
 
-
-            correo.enviarCorreoNota(matricula.getAlumnoEntity().getPersonaEntity().getPersona_nombre(), asignaturaDocente.getAsignaturaEntity().getAsignatura_nombre(), notaRequest.getNota(), asignaturaDocente.getDocenteCursoEntity().getDocenteEntity().getPersonaEntity().getPersona_nombre());
+       //     correo.enviarCorreoNota(matricula.getAlumnoEntity().getPersonaEntity().getPersona_nombre(), asignaturaDocente.getAsignaturaEntity().getAsignatura_nombre(), notaRequest.getNota(), "test", new Date().toString());
 
             return ResponseEntity.ok(new MessageResponse("Se ha registrado la nota con éxito!"));
-        } catch (Exception e) {
-            return ResponseEntity.badRequest().body(new MessageResponse("Error: No se ha logrado registrar la nota!" + e.getMessage()));
-        }*/
-
-        LocalDate localDate = notaRequest.getFecha().toInstant().atZone(ZoneId.systemDefault()).toLocalDate();
-        int year = localDate.getYear();
-
-        MatriculaEntity matricula = matriculaService.findMatriculaByRunAndCurso(notaRequest.getRun(), year, notaRequest.getEstablecimiento());
-        if (matricula == null) {
-            return ResponseEntity.badRequest().body(new MessageResponse("Error: No se ha logrado encontrar matricula!"));
         }
 
-        CursoEstablecimientoEntity cursoEstablecimiento = cursoEstablecimientoService.findById(matricula.getCursoEstablecimientoEntity().getCurso_establ_id());
-        if (cursoEstablecimiento == null) {
-            return ResponseEntity.badRequest().body(new MessageResponse("Error: No se ha logrado encontrar matricula !"));
-        }
+        return ResponseEntity.badRequest().body(new MessageResponse("Error: No se ha logrado registrar la nota!" ));
 
-        AsignaturaDocenteEntity asignaturaDocente = asignaturaDocenteService.findDocenteCursoByRunAndAsignaturaAndEstablecimiento(notaRequest.getDocente_run(), notaRequest.getAsignatura(), cursoEstablecimiento.getCursoEntity().getCurso_id(), year);
-        if (asignaturaDocente == null) {
-            return ResponseEntity.badRequest().body(new MessageResponse("Error: No se ha logrado registrar nota!"));
-        }
 
-        // Crear nueva entidad AsignaturaNotaEntity
-        AsignaturaNotaEntity nota = new AsignaturaNotaEntity();
-        nota.setMatriculaEntity(matricula);
-        nota.setAsignaturaDocenteEntity(asignaturaDocente);
-        nota.setAsignatura_nota_nota(notaRequest.getNota());
-        nota.setPosicion_nota_nota(notaRequest.getPosicion_nota());
-        // Guardar la nueva entidad en la base de datos
-        asignaturaNotaService.save(nota);
-
-        correo.enviarCorreoNota(matricula.getAlumnoEntity().getPersonaEntity().getPersona_nombre(), asignaturaDocente.getAsignaturaEntity().getAsignatura_nombre(), notaRequest.getNota(), asignaturaDocente.getDocenteCursoEntity().getDocenteEntity().getPersonaEntity().getPersona_nombre(), new Date().toString());
-
-        return ResponseEntity.ok(new MessageResponse("Se ha registrado la nota con éxito!"));
     }
-
-
     @PutMapping("/Update")
     public ResponseEntity<?> editarNota(@Valid @RequestBody EditNotas notaRequest) {
 
         try{
-            // Editar  entidad AnotacionesEntity
-            AsignaturaNotaEntity nota = asignaturaNotaService.findById(notaRequest.getAsignatura_nota_id());
-            nota.setAsignatura_nota_nota(notaRequest.getNota());
+            if(notaRequest.IsValid()){
+                // Editar  entidad AnotacionesEntity
+                AsignaturaNotaEntity nota = asignaturaNotaService.findById(notaRequest.getAsignatura_nota_id());
+                nota.setAsignatura_nota_nota(notaRequest.getNota());
 
-            // Guardar la nueva entidad en la base de datos
-            asignaturaNotaService.save(nota);
+                // Guardar la nueva entidad en la base de datos
+                asignaturaNotaService.save(nota);
 
-            return ResponseEntity.ok(new MessageResponse("Se ha editado nota con exito!"));
+                return ResponseEntity.ok(new MessageResponse("Se ha editado nota con exito!"));
+            }
 
+            return  ResponseEntity.badRequest().body(new MessageResponse("Error: No se ha logrado editar nota!"));
 
         }catch (Exception e){
 
@@ -147,14 +143,16 @@ public class NotasController {
     public ResponseEntity<?> deleteNota(@Valid @RequestBody EditNotas notaRequest) {
 
         try{
+            if(notaRequest.IsValid()){
+                // Eliminar nueva entidad AnotacionesEntity
+                AsignaturaNotaEntity nota = asignaturaNotaService.findById(notaRequest.getAsignatura_nota_id());
+                // Guardar la nueva entidad en la base de datos
+                asignaturaNotaService.delete(nota);
 
-            // Eliminar nueva entidad AnotacionesEntity
-            AsignaturaNotaEntity nota = asignaturaNotaService.findById(notaRequest.getAsignatura_nota_id());
-            // Guardar la nueva entidad en la base de datos
-            asignaturaNotaService.delete(nota);
+                return ResponseEntity.ok(new MessageResponse("Se ha eliminado nota con exito!"));
+            }
 
-            return ResponseEntity.ok(new MessageResponse("Se ha eliminado nota con exito!"));
-
+            return  ResponseEntity.badRequest().body(new MessageResponse("Error: No se ha logrado eliminar nota!"));
 
         }catch (Exception e){
 
@@ -162,4 +160,5 @@ public class NotasController {
 
         }
     }
+
 }
